Update and rename planet_orbits.py to planet_game.py
Added buttons that allows user to add a celestial object to the solar system. This object will enter at a straight line as if a foreign object was passing through the solar system. Still need to add more functionality to make it feel like a game, such as a score counter for collisions.